+++ app/routes/howTo.jsx
@@ -1,0 +1,12 @@
+import Header from "../components/header";
+
+function howTo() {
+  return (
+    <>
+      <Header />
+      <h1>アプリの使い方</h1>;
+    </>
+  );
+}
+
+export default howTo;
