+++ app/styles/auth/gamen.css
@@ -1,3 +1,4 @@
-h1{
-    padding-top: 40px;
+.contentsWrapper {
+  margin: 0, auto;
+  padding-top: 72px;
 }
